fix(core): correctly compute pipeline graph scroll position
